requirements.txt
flask
gunicorn
google-cloud-storage
pydantic
google-genai
google-cloud-aiplatform
google-adk
